requirements.txt
openai-whisper
sounddevice
numpy
wavio
webrtcvad
huggingface_hub
requests
spacy
edge-tts
